begin refactorng of resource and artifact spec
Remove Artifact and Resource classes and use these names
for the ArtifactSpec and ResourceSpec.

They have two new abstract method required:

* `Artifact.resolve`: finds and fetch the artifact data for the current
revision. This will be used to get the artifact that the strategy will
deploy and avoid having this logic duplicated in steps
(see docker_pull).
* `Resource.state`: gets the current state for the external resource.
Similarly to the above method, this a common operation when dealing
with resources and should not be duplicated.

The new `Artifact` has also a method `exists` that uses `resolve` for
its default implementation, but better logic should be used when
possible (see `DockerImage`).